+++ Tests/Wire/Resources/Shipping/OrderDrafts/TaskGroup/TaskGroupClientWireTests.swift
@@ -53,6 +53,29 @@ import Chrt
         try #require(response == expectedResponse)
     }
 
+    @Test func setVehicleTypeV11() async throws -> Void {
+        let stub = HTTPStub()
+        stub.setResponse(
+            body: Data(
+                """
+                true
+                """.utf8
+            )
+        )
+        let client = ChrtClient(
+            baseURL: "https://api.fern.com",
+            token: "<token>",
+            urlSession: stub.urlSession
+        )
+        let expectedResponse = true
+        let response = try await client.shipping.orderDrafts.taskGroup.setVehicleTypeV1(
+            taskGroupId: "task_group_id",
+            request: .init(),
+            requestOptions: RequestOptions(additionalHeaders: stub.headers)
+        )
+        try #require(response == expectedResponse)
+    }
+
     @Test func setTaskOrderingV11() async throws -> Void {
         let stub = HTTPStub()
         stub.setResponse(