+++ Tests/Wire/Resources/Shipping/Tasks/TasksClientWireTests.swift
@@ -118,6 +118,7 @@ import Chrt
         let expectedResponse = true
         let response = try await client.shipping.tasks.attemptV1(
             taskId: "task_id",
+            request: .init(),
             requestOptions: RequestOptions(additionalHeaders: stub.headers)
         )
         try #require(response == expectedResponse)